Move new screenshot test to test_qt_mainwindow.py
