+++ napari/_qt/_tests/test_qt_viewer.py
@@ -8,11 +8,10 @@ from unittest import mock
 import numpy as np
 import pytest
 from imageio import imread
-from qtpy.QtGui import QGuiApplication, QImage
+from qtpy.QtGui import QGuiApplication
 from qtpy.QtWidgets import QMessageBox
 
 from napari._qt.qt_viewer import QtViewer
-from napari._qt.utils import QImg2array
 from napari._tests.utils import (
     add_layer_by_type,
     check_viewer_functioning,
@@ -297,43 +296,6 @@ def test_screenshot_dialog(make_napari_viewer, tmpdir):
     output_data = imread(expected_filepath)
     expected_data = viewer.window._qt_viewer.screenshot(flash=False)
     assert np.allclose(output_data, expected_data)
-
-
-@skip_on_win_ci
-def test_screenshot_to_file(make_napari_viewer, tmp_path):
-    """Test taking a screenshot and saving it to a file."""
-    viewer = make_napari_viewer()
-    screenshot_dir_path = tmp_path / "screenshots"
-    screenshot_dir_path.mkdir()
-    screenshot_file_path = str(screenshot_dir_path / "screenshot.png")
-
-    np.random.seed(0)
-    # Add image
-    data = np.random.random((10, 15))
-    viewer.add_image(data)
-
-    # Add labels
-    data = np.random.randint(20, size=(10, 15))
-    viewer.add_labels(data)
-
-    # Add points
-    data = 20 * np.random.random((10, 2))
-    viewer.add_points(data)
-
-    # Add vectors
-    data = 20 * np.random.random((10, 2, 2))
-    viewer.add_vectors(data)
-
-    # Add shapes
-    data = 20 * np.random.random((10, 4, 2))
-    viewer.add_shapes(data)
-
-    # Take screenshot
-    screenshot_array = viewer.window.screenshot(
-        screenshot_file_path, flash=False, canvas_only=True
-    )
-    screenshot_array_from_file = QImg2array(QImage(screenshot_file_path))
-    assert np.array_equal(screenshot_array, screenshot_array_from_file)
 
 
 @pytest.mark.parametrize(
